Final version

+++ src/core/database.rs
@@ -2,12 +2,20 @@ use rusqlite::Connection;
 use crate::models::error::FurbrowserResult;
 use crate::util::sql::SyncSQLFetch;
 
-pub fn open_database(file_path: &str) -> FurbrowserResult<Connection> {
+pub fn open_database(file_path: &str, backward_compatibility: bool) -> FurbrowserResult<Connection> {
     println!("Opening database...");
     let connection = Connection::open(file_path)?;
-    connection.execute("DROP TABLE IF EXISTS tags", ())?;
+
+    if !backward_compatibility {
+        connection.execute("DROP TABLE IF EXISTS tags", ())?;
+    }
+
     connection.execute("CREATE TABLE IF NOT EXISTS images (id INT NOT NULL, liked BOOL, disliked BOOL, PRIMARY KEY (id))", ())?;
-    connection.execute("CREATE TABLE IF NOT EXISTS published (id INT UNIQUE NOT NULL, processed BOOL NOT NULL, vote BOOL NOT NULL, PRIMARY KEY (id))", ())?;
+
+    if backward_compatibility {
+        connection.execute("CREATE TABLE IF NOT EXISTS published (id INT UNIQUE NOT NULL, processed BOOL NOT NULL, vote BOOL NOT NULL, PRIMARY KEY (id))", ())?;
+    }
+
     Ok(connection)
 }
